requirements.txt
bcrypt==3.2.0
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Jinja2==3.0.1
Pillow==8.3.1
SQLAlchemy==1.4.22
WTForms==2.3.3
pandas_alive==0.2.4
pandas==1.3.1
bar_chart_race==0.1.0
IPython == 7.26.0